Replace factory functions with classes.
Element() and ElementTree() were factory functions until now.
They are now defined as classes.

With this change, we can annotate code with the public Element class,
instead of using the private _Element class.
For example:
```
def print_tree(tree: e.ElementTree[e.Element]) -> None: ...
```

There is a related PR that was merged into lxml:
https://github.com/lxml/lxml/pull/405

In the lxml PR, _Element is a virtual subclass of Element.
This means that when we create an Element it actually creates an _Element class,
but the created object is also an instance of Element:
```
el = Element()
reveal_type(el)  # _Element
isinstance(element, Element)  # True
isinstance(element, _Element)  # True
```